Move File RegisterView
